Extract some reducer helpers
